Model training, grid search, model loading/saving
Make sure to check these variables to get the desired functionality:
load
dump
testSVM

Still running GridSearch so model isn't optimised. Additionally, the bones of loading the different batches as scaled data is present, but not intergrated just yet. May not be necessary in the final build, but I was bored. This is mainly to d o the fitting in increments to avoid memory / computation issues